place and room page

+++ app/src/main/java/com/example/learning_android/domain/model/Device.kt
@@ -5,7 +5,7 @@ import java.time.Instant
 data class Device(
     val id: String,
     val name: String,
-    val batteryPercentage: Float?,
+    val batteryPercentage: Int?,
     val batteryReadAt: Instant?,
     val placeId: String?,
     val type: DeviceType
